+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* Fixed error for partial result option if field contains box.NULL.
+
 ## [0.6.0] - 2021-03-29
 
 ### Fixed